+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 0.6.9-dev01
+- ref: remove unused view-state and view-model
 - feat: fill storage internal/external functionality
 - feat: storage detail and fill text field & options
 - feat: storage capacity MB/GB filter